requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
colorama>=0.4.6
rich>=13.0.0
typing-extensions>=4.8.0 